+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Adding S3 Origins
 - Adding default_root_object
+- Adding lambda at edge associations
 - Updating some default variable values
 
 ## v1.0.0